# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/achievements/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Получение достижений",
                "parameters": [
                    {
                        "type": "string",
                        "name": "account_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseWithDetails-handler_listAchievementResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/achievements/{AccountId}/{AchievementId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Получение достижения",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор достижения",
                        "name": "AchievementId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор аккаунта",
                        "name": "AccountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseWithDetails-handler_getAchievementResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/quests/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Получение квестов",
                "parameters": [
                    {
                        "type": "string",
                        "name": "settlement_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseWithDetails-handler_listQuestsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/quests/{questId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Получение информации о квесте",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор квеста",
                        "name": "questId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseWithDetails-handler_getQuestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/settlements/search": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settlements"
                ],
                "summary": "Поиск города",
                "parameters": [
                    {
                        "type": "string",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseWithDetails-handler_searchSettlementsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.QuestType": {
            "type": "string",
            "enum": [
                "route"
            ],
            "x-enum-varnames": [
                "RouteQuestType"
            ]
        },
        "handler.getAchievementResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "passed": {
                    "type": "integer"
                },
                "steps": {
                    "type": "integer"
                }
            }
        },
...
        "handler.getQuestResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "preview": {
                    "$ref": "#/definitions/handler.image"
                },
                "reward": {
                    "type": "number"
                },
                "settlement_id": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.getQuestResponseStep"
                    }
                },
                "type": {
                    "$ref": "#/definitions/entity.QuestType"
                }
            }
        },
        "handler.getQuestResponseStep": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.image"
                    }
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "place_type": {
                    "type": "string"
                },
                "quest_id": {
                    "type": "string"
                },
                "schedule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.getQuestResponseStepSchedule"
                    }
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "handler.getQuestResponseStepSchedule": {
            "type": "object",
            "properties": {
                "from_to": {
                    "type": "string"
                },
                "week_day": {
                    "type": "string"
                }
            }
        },
        "handler.image": {
            "type": "object",
            "properties": {
                "sizes": {
                    "$ref": "#/definitions/handler.sizes"
                }
            }
        },
        "handler.listAchievementResponseElement": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "passed": {
                    "type": "integer"
                },
                "steps": {
                    "type": "integer"
                }
            }
        },
...
        "handler.listQuestsResponseElement": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "preview": {
                    "$ref": "#/definitions/handler.image"
                },
                "reward": {
                    "type": "number"
                },
                "settlement_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/entity.QuestType"
                }
            }
        },
        "handler.searchSettlementsResponseElement": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.size": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.sizes": {
            "type": "object",
            "properties": {
                "m": {
                    "$ref": "#/definitions/handler.size"
                },
                "o": {
                    "$ref": "#/definitions/handler.size"
                },
                "x": {
                    "$ref": "#/definitions/handler.size"
                }
            }
        },
        "shttp.ResponseError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {}
            }
        },
...
        "shttp.ResponseWithDetails-handler_getAchievementResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/handler.getAchievementResponse"
                }
            }
        },
...
        "shttp.ResponseWithDetails-handler_getQuestResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/handler.getQuestResponse"
                }
            }
        },
        "shttp.ResponseWithDetails-handler_listAchievementResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.listAchievementResponseElement"
                    }
                }
            }
        },
...
        "shttp.ResponseWithDetails-handler_listQuestsResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.listQuestsResponseElement"
                    }
                }
            }
        },
        "shttp.ResponseWithDetails-handler_searchSettlementsResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.searchSettlementsResponseElement"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ExternalId": {
            "type": "apiKey",
            "name": "ExternalId",
            "in": "header"
        }
    }
}